.qodo/backend/requirements.txt
fastapi
sqlalchemy
bcrypt
python-dotenv
uvicorn
openai-agents
httpx 